requirements.txt
python-telegram-bot==22.0
flask==2.3.3
requests==2.31.0 